+++ backend/packages/integrations/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/integrations
 
+## 0.21.5
+
+### Patch Changes
+
+- Updated dependencies [ea59e91]
+  - @cat-factory/contracts@0.30.0
+  - @cat-factory/kernel@0.33.0
+
 ## 0.21.4
 
 ### Patch Changes